// src/app/shared/enum/starting-price-criteria.enum.ts
export enum StartingPriceCriteria {
  PRICELIST1 = 'PRICELIST1',
  PRICELIST2 = 'PRICELIST2',
  PRICELIST3 = 'PRICELIST3',
  PRICELIST4 = 'PRICELIST4',
  PRICE = 'PRICE',
  PRICE_OF_SALE = 'PRICE_OF_SALE',
  VALUE_PRICE_LAST_SALE = 'VALUE_PRICE_LAST_SALE',
  VALUE_PRICE_AVERAGE_SALE = 'VALUE_PRICE_AVERAGE_SALE',
  VALUE_PRICE_AVERAGE_SALE_MOV = 'VALUE_PRICE_AVERAGE_SALE_MOV',
  VALUE_PRICE_LAST_PURCH = 'VALUE_PRICE_LAST_PURCH',
  VALUE_PRICE_AVERAGE_PURCH = 'VALUE_PRICE_AVERAGE_PURCH',
  VALUE_PRICE_LOW_PURCH = 'VALUE_PRICE_LOW_PURCH',
  VALUE_PRICE_MAX_PURCH = 'VALUE_PRICE_MAX_PURCH',
}