// src/app/shared/enum/recharge-discharge-criteria.enum.ts
export enum RechargeDischargeCriteria {
  NOTHING = 'NOTHING',
  RECHARGE = 'RECHARGE',
  RECHARGE_SALE = 'RECHARGE_SALE',
  CUSTOM_RECHARGE_PERCENTAGE = 'CUSTOM_RECHARGE_PERCENTAGE',
  CUSTOM_RECHARGE_AMOUNT = 'CUSTOM_RECHARGE_AMOUNT',
  CUSTOM_DESCHARGE_PERCENTALE = 'CUSTOM_DESCHARGE_PERCENTALE',
  CUSTOM_DESCHARGE_AMOUNT = 'CUSTOM_DESCHARGE_AMOUNT',
  PROPOSE_DESCHARGE_FORMULA = 'PROPOSE_DESCHARGE_FORMULA'
}